+++ src/random.hpp
@@ -38,7 +38,7 @@ public:
 
   uint32_t generate () {
     next ();
-    return state >> 32;
+    return static_cast <uint32_t> (state >> 32);
   }
   int generate_int () { return (int) generate (); }
   bool generate_bool () { return generate () < 2147483648u; }
